import type and change where we do the map
