+++ ui/desktop/src/components/ProgressiveMessageList.tsx
@@ -19,7 +19,7 @@ import { Message } from '../api';
 import GooseMessage from './GooseMessage';
 import UserMessage from './UserMessage';
 import { SystemNotificationInline } from './context_management/SystemNotificationInline';
-import { NotificationEvent } from '../hooks/useChatStream';
+import { NotificationEvent } from '../types/message';
 import LoadingGoose from './LoadingGoose';
 import { ChatType } from '../types/chat';
 import { identifyConsecutiveToolCalls, isInChain } from '../utils/toolCallChaining';
