+++ ui/desktop/src/components/ToolCallWithResponse.tsx
@@ -4,10 +4,13 @@ import React, { useEffect, useRef, useState } from 'react';
 import { Button } from './ui/button';
 import { ToolCallArguments, ToolCallArgumentValue } from './ToolCallArguments';
 import MarkdownContent from './MarkdownContent';
-import { ToolRequestMessageContent, ToolResponseMessageContent } from '../types/message';
+import {
+  ToolRequestMessageContent,
+  ToolResponseMessageContent,
+  NotificationEvent,
+} from '../types/message';
 import { cn, snakeToTitleCase } from '../utils';
 import { LoadingStatus } from './ui/Dot';
-import { NotificationEvent } from '../hooks/useChatStream';
 import { ChevronRight, FlaskConical } from 'lucide-react';
 import { TooltipWrapper } from './settings/providers/subcomponents/buttons/TooltipWrapper';
 import MCPUIResourceRenderer from './MCPUIResourceRenderer';
